requirements.txt
asgiref==3.4.1
astroid==2.8.3
autopep8==1.5.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-ajax-selects==2.1.0
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-js-asset==1.2.2
django-taggit==1.5.1
django-taggit-serializer==0.1.7
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
environs==9.3.4
idna==3.3
isort==5.9.3
itypes==1.2.0
Jinja2==3.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
marshmallow==3.14.0
mccabe==0.6.1
oauthlib==3.1.1
pep8==1.7.1
Pillow==8.4.0
platformdirs==2.4.0
pycodestyle==2.8.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.2
